rtc-sse/requirements.txt
Django==5.0.6
environs[django]==11.0.0
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-eventstream==5.1.0
daphne==4.1.2